Moving preference management to link component and flagging as dismissed when generating background
