+++ plugins/woo-ai/src/image-background-removal/background-removal-link.tsx
@@ -3,9 +3,11 @@
  */
 import { useState, useEffect } from 'react';
 import { __experimentalUseBackgroundRemoval as useBackgroundRemoval } from '@woocommerce/ai';
+import { store as preferencesStore } from '@wordpress/preferences';
 import { __ } from '@wordpress/i18n';
 import { createInterpolateElement } from '@wordpress/element';
 import { Notice } from '@wordpress/components';
+import { useDispatch, select } from '@wordpress/data';
 
 /**
  * Internal dependencies
@@ -19,6 +21,8 @@ import {
 	getCurrentAttachmentDetails,
 } from './image_utils';
 import { TourSpotlight } from '../components/';
+
+const preferenceId = `spotlightDismissed-backgroundRemovalLink`;
 
 const getErrorMessage = ( errorCode?: string ) => {
 	switch ( errorCode ) {
@@ -44,6 +48,11 @@ const recordBgRemovalTracks = recordTracksFactory(
 export const BackgroundRemovalLink = () => {
 	const { fetchImage } = useBackgroundRemoval();
 	const { showSnackbar, removeSnackbar } = useFeedbackSnackbar();
+	const hasBeenDismissedBefore = select( preferencesStore ).get(
+		'woo-ai-plugin',
+		preferenceId
+	);
+	const { set } = useDispatch( preferencesStore );
 
 	const [ state, setState ] = useState< 'none' | 'generating' | 'uploading' >(
 		'none'
@@ -53,6 +62,9 @@ export const BackgroundRemovalLink = () => {
 	useEffect( () => {
 		recordBgRemovalTracks( 'view_ui' );
 	}, [] );
+
+	const setSpotlightAsDismissed = () =>
+		set( 'woo-ai-plugin', preferenceId, true );
 
 	const onRemoveBackgroundClick = async () => {
 		removeSnackbar();
@@ -89,6 +101,8 @@ export const BackgroundRemovalLink = () => {
 			} );
 
 			recordBgRemovalTracks( 'complete' );
+
+			setSpotlightAsDismissed();
 			showSnackbar( {
 				label: __( 'Was the generated image helpful?', 'woocommerce' ),
 				onPositiveResponse: () => {
@@ -137,39 +151,42 @@ export const BackgroundRemovalLink = () => {
 				</button>
 				<img src={ MagicIcon } alt="" />
 			</div>
-			<TourSpotlight
-				id="backgroundRemovalLink"
-				reference={ `#${ LINK_CONTAINER_ID }` }
-				description={ __(
-					'Effortlessly make your product images pop by removing the background using state-of-the-art AI technology. Just click the button and watch!',
-					'woocommerce'
-				) }
-				title={ createInterpolateElement(
-					__(
-						'<NewBlock /> Remove backgrounds with AI',
+			{ ! hasBeenDismissedBefore && (
+				<TourSpotlight
+					id="backgroundRemovalLink"
+					reference={ `#${ LINK_CONTAINER_ID }` }
+					description={ __(
+						'Effortlessly make your product images pop by removing the background using state-of-the-art AI technology. Just click the button and watch!',
 						'woocommerce'
-					),
-					{
-						NewBlock: (
-							<span className="woo-ai-background-removal-link__new-block">
-								NEW
-							</span>
+					) }
+					title={ createInterpolateElement(
+						__(
+							'<NewBlock /> Remove backgrounds with AI',
+							'woocommerce'
 						),
+						{
+							NewBlock: (
+								<span className="woo-ai-background-removal-link__new-block">
+									{ __( 'NEW', 'woocommerce' ) }
+								</span>
+							),
+						}
+					) }
+					placement="left"
+					spotlightParent={
+						( document.querySelector(
+							`#${ LINK_CONTAINER_ID }`
+						) as HTMLElement ) ?? document.body
 					}
-				) }
-				placement="left"
-				spotlightParent={
-					( document.querySelector(
-						`#${ LINK_CONTAINER_ID }`
-					) as HTMLElement ) ?? document.body
-				}
-				onDismissal={ () =>
-					recordBgRemovalTracks( 'spotlight_dismissed' )
-				}
-				onDisplayed={ () =>
-					recordBgRemovalTracks( 'spotlight_displayed' )
-				}
-			/>
+					onDismissal={ () => {
+						recordBgRemovalTracks( 'spotlight_dismissed' );
+						setSpotlightAsDismissed();
+					} }
+					onDisplayed={ () =>
+						recordBgRemovalTracks( 'spotlight_displayed' )
+					}
+				/>
+			) }
 			{ displayError && (
 				<Notice onRemove={ () => setDisplayError( null ) }>
 					{ displayError }
